requirements.txt
# AgentScope Framework
agentscope>=1.0.0

# Document Processing
pypdf>=5.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11

# Academic API Clients
requests>=2.28.0
arxiv>=1.4.0
scholarly>=1.7.0

# Data Processing
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.64.0
colorama>=0.4.5
